metadata_generator: move tool shed impl into tool shed package
the correct home and allows better typing with actual Repository type annotations
